About there on the schema front, messy solution!
